Add a test framework
Add a unit test case.  This is really messy, because I was unable to make the test work in all circumstances.  With the egg installed in site-packages the imports get all weirded up when running nosetests, and there were also cases where running the test from the command line wasn't working.  I added exceptions so that the test will only run from nosetests with the egg not present in site-packages, but there's got to be a better way.