Break plugin dependency on Snarl
Add a builder method and set that as the nose entry point so that the plugin code itself does not need to import anything (like snarler) that may itself import PySnarl.  Since I don't have PySnarl on my Mac, this lets me run the unit tests against the plugin (the thing I care most about) regardless of the presence of PySnarl.