Refactor plugin, make real snarler match test double
Get finalize down to three lines, move conditional out to a more appropriate method, break out formatting of time delta from method that concatenates the elapsed time string, make the ternary in _test_summary a little more clear, gun options and configure as they just call through to super and are therefore clutter (after confirming that I still get autogeneration of the --help string and the --with-snore parameters).  Update snarler.py to match the interface evolved in the test double contained within test_plugin.py so that the egg will install and run.