Fix a bug found on first Windows run
pkg_resources.resource_filename constructs proper paths using proper directory separators on both Windows and OS X, it seems.  I had a hardcoded '/' when building the icon path that caused three unit tests to fail the first time I tried to run them on Windows, so I removed the hardcoded '/' and instead use the '/' within the resource_filename call so that it is automatically converted to '\' on Windows and the tests pass.